pkg: objectstate: rte: embed reference state
The original intent of `objectstate` packageswas to gather in one set
of data structures (with consistent APIs) the current and desired
state of relevant manifests.
The intended flow was like
```
...
existingManifests := objectstate.FromClient(...)

for _, object := existingManifests.State(desiredManifests) {
	// compare object.Existing with object.Desired
	// using object.Compare, merge changes using object.Merge
	...
}
```

`FromClient` may need extra parameters, like the cluster platform
which need to be consistent with the same parameters passed to `State`.

Up until now, for `objectstate/rte` the two functions just required
a (large) set of parameters, trusting the caller to be consistent.
This was mostly a historical artifact rather than a deliberate decision.

With this commit, we make things consistent by recording the parameters
got from `FromClient`. Now `State` ismuch lighter.
This improves consistency and streamlines the UX.

Should the client need to change the objectstate parameters, it should
create a new instance with `FromClient`, and this was actually always
the intended usage. It is more evident now, which is a extra benefit.

Signed-off-by: Francesco Romani <fromani@redhat.com>