objectstate: objectupdate: rte: shuffle updates
We should not do any update in objectstate code
if we can help it - this is why we have objectupdate and
the orchestration code in the controllers.

Again, updates were done here mostly because of historical
accident. This is a chance to clean it up: move update
in a callback, with no intended changes in behaviour.

Signed-off-by: Francesco Romani <fromani@redhat.com>