e2e:serial: check scheduling error per topology scope
When the scheduler plugin failed to align a pod into a node it reports an error.
The error message is varies upon the  configuration.

We should support a check of the scheduling error for both container and pod scope configurations.
This patch adds the support for the container scope (pod scope support is already exists)

Signed-off-by: Talor Itzhak <titzhak@redhat.com>