e2e:serial: place pods only on multi NUMA nodes
There are times when single numa machines are in the cluster (usually developer use-case).
By default, if the machine is a worker, the daemonset will try to place a pod in there.

Usually, single NUMA machines have fewer resources and they are getting filtered by the `NodeResourceFit` plugin, hence they'll never reach the `NodeTopologyMatch` filtering process, which in turn leads to test failure.

This is because the test expects the pod to be rejected by the `NodeTopologyMatch` plugin and it's looking for the plugin's error message, but unable to find it.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>