controller: drop deployer helper
We thought we could make use of the deployer Helper construct.
This turned out to be a poor abstraction which we use exactly once.
So let's stop using it, which will enable the deployer package to
provide better alternatives in the future (less clients -> more freedom
to change).

No intended changes in behaviour.

Signed-off-by: Francesco Romani <fromani@redhat.com>