version: minimize before to store
The full long version MAJOR.MINOR.PATCH-RELEASE.EXTRA
is going to be misdetected by the deployer code (4.11.0-0 is going to
fail the AtLeastVersionString check), so let's strip everything
except MAJOR.MINOR.PATCH before to store the detected version.
This is less precise but still compliant with the intent of the code,
and should still be precise enough for our needs.

Signed-off-by: Francesco Romani <fromani@redhat.com>