e2e: install: properly wait when removing NRO objs
We have a install test which wants to delete and recreate
a NRO object. The concept is fine per se, but deleting and
creating NRO means applying Machine Configs on OCP - in both
cases, and this operation is both a separate step, which should
be checked individually, and time expensive.

So we extract the wait helpers and we call them in this flow.

This way we have better control over the test flow and we
can pinpoint more precisely what is failing and where.

Signed-off-by: Francesco Romani <fromani@redhat.com>