e2e: serial: handle > 2 workers
In case we have more than two workers, but some of them have
inconsistent NUMA zones (e.g. BM + VM workers in the same pool),
we can have a panic due to the void assumptions that all the workers
have the same NUMA zones.

This is arguably more a cluster configuration issue than a test flaw,
but the test should not panic nevertheless.

Signed-off-by: Francesco Romani <fromani@redhat.com>