render: add option to render only the CRD
In the regular and expected flows, the operator will own and manage the
API CRD when asked to deploy the main part (NRO object, deploying API +
RTEs). In some cases, most notably to test partial scenarios or
non-regressions, is it handy to have the CRD definition alone, to easily
set up scenarios.

We evaluated adding a trivial, 20-30 lines new helper tool just to emit
the CRD manifests, but while the source code was trivial, the binary
artifact was a whopping 30 megs (k8s deps), so the best compromise
seems to extend the existing render mode of the operator.

Signed-off-by: Francesco Romani <fromani@redhat.com>