e2e: serial: move wait function in fixture pkg
In order to prepare the code to move the `objects/wait` util pkg
in the `internal` hierarchy, trivial code movement of the
`ForPaddingPods` function from `objects/wait` to `fixture` pkg.
This is because this function is the only one which depends on
code not yet moving to `internal`. IOW, it depends on code
which is still kept in `test/utils`, so it should stay in `test/utils`.

Signed-off-by: Francesco Romani <fromani@redhat.com>