serial: stabilize pod requests and free resources of cpu test
The test currently requests 9 cpus in total and on the target node expects to leave room for 8 cpus. Although we don't anticipate the pod to be scheduled we should make the target node feasible on compute level to accept the pod. The current placement does not have a problem because baseload of cpus is rounded up to 2 cpus, yet we should consider having the node feasible regardless of the baseload. see inline comments for more detials.

Signed-off-by: shereenH <shajmakh@redhat.com>