serial: fix waiting for pod phase
The current use of wait.ForPodPhase() is not relevant for this place in the test because it continues to next steps once the pod reached the desired phase. What we want to check in fact is that the pod keeps on pending for some time, this functionality is provided by using wait.WhileInPodPhase().

Signed-off-by: shereenH <shajmakh@redhat.com>