e2e: serial: nonreg: fix target node selection
tests should not assume any worker in the node group (let alone in the
cluster) is suitable as target, and should pick it among the subset of
nodes the test setup identified as test candidate.

This is first and foremost to make the test prerequisite and intentions
clearer, but also to deal with mixed HW CI environments.

Signed-off-by: Francesco Romani <fromani@redhat.com>