e2e: serial: don't use pause on initcnt
We should never use `pause` as entrypoint in initcnt, because
`pause` by design will never terminate, so the init containers
will never terminate, so the pod will never go running.

Now that we have an enhanced pause image (or a image with both
pause and sleep in general), we can safely change the entrypoint
to just sleep for a tiny bit as entry point.

Signed-off-by: Francesco Romani <fromani@redhat.com>