e2e: fixture: serial: configurable cooldown
We originally added the cooldown time to allow the cluster state
to settle down after a test run. But that was always meant a safeguard.
Now we have 50+ specs, and 50+ specs each waiting 30s is now
contributing significantly to the suite wall clock execution time.

Let's start to improve. Let's add a env var to override this value,
so we can start checking if lower values (ideally zero) are now OK.

Signed-off-by: Francesco Romani <fromani@redhat.com>