e2e:serial:tmpolicy: test device NUMA alignment
Schedule a pod with 2 containers, where the containers` requests are available under two possible nodes, but only one of them has the devices on a single numa in such a way that each device is available on a different numa.
The pod is expected to be scheduled on the feasible node with relevant conditions in terms of availability on single numa nodes.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>