e2e:serial:tmpolicy: filtering function for scope=container
When scope=container we should filter the matching nodes based on the total amount of resources they have (and not per zone as we did so far).
The reason for that is because not all the containers have to fit into a single zone, but can be spreads across different zones.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>