serial: set container images via the function
Some CI environments are disconnected and cannot pull images outside the local host image registry. A way to overcome this unreachability is via setting one of the environment variables: E2E_NROP_URL_PAUSE_IMAGE, E2E_PAUSE_IMAGE_URL to have an image that is reachable by the local host.
Thus tests should not use the value of PauseImage directly but via the function that fetches the desired image to use for the tests which is images.GetPauseImage().

Signed-off-by: shereenH <shajmakh@redhat.com>