utils: remove pause image duplicated const
The images utility already has the constant of PauseImage which is the actual one being used by tests containers. Move the PauseCommand to images utility as well to align related conts in one file for better mainatiance and to avoid using objects.PauseImage directly which is not recommended for disconnected environments. see this commit for more details:
https://github.com/openshift-kni/numaresources-operator/pull/417/commits/2cd1fcd40c1dfe8e6ba0077332057b1f61fc902f

Signed-off-by: shereenH <shajmakh@redhat.com>