e2e: serial: fixture: configurable teardown time
We continue our quest to cut down the test execution wall clock time
begun in 3aed1168f2684f29276bc308890de319e72046eb .

One of the reasons why we introduced the cooldown at all, was
to make sure the cluster was settled after the test execution.
But this should actually be done at test teardown time.

The test teardown must ensure the cluster is restore at its state
before the current test started, or it's a bug in the test fixture.

In practical terms, we add another knob to configure the test
*teardown* time, so we can play with lower cooldown but increase
the teardown if needed, to improve the execution time without
destabilizing (read: uncover too many issues at once) the suite.

The key difference is we always wait for the full cooldown time,
inconditionally, while we wait UP TO the teardown time.
So lower (ideally zero) cooldown but higher teardown time should
yield both a more robust and faster test suite.

Signed-off-by: Francesco Romani <fromani@redhat.com>