serial: remove pod scope padding function
Till now, the padding was done with the assumption that the pod's containers are all expected to land on one numa zone (the normal anticipated behavior). So for the target node, we were padding the first numa "fully" but leaving minimal free resources (2 cpus and 2Gi memory). while padding the second numa (on the target node) leaving space to total resources of all the pod's containers. This however is not always the padding we want to do, especially when we are verifying negative checks, for example the total pod's resources are available but on multi numas.

To manage the padding for negative checks we better go with the way of padding for container scope tests (e.g free resources per each numa).

For the positive tests, we'll still be able to apply the same old padding but now by specifying the total resources in one numa as free resource on target node, and unsuitable (minimal) resources on the other numa. (If we specify on one numa only the other numa won't be padded)