serial: add negative test for pod scope
Add negative test that schedules a pod with two cont each requesting a different device type. The resources are available on one node but on different single numas in such a way that if the scope was container the pod will successfully be scheduled on that node. The pod is expected to keep on pending.

Signed-off-by: shereenH <shajmakh@redhat.com>