rte: finer grained node group configuration
Introduce per-node-group config (NodeGroupConfig) object,
in order to enable fine-tuning of RTE.

This is needed to enable cluster-admins to explore the proper
tradeoff between cluster load and scheduling precision.

We will also need to have some slack to explore the
best default to set, and we need some scalability knobs
for the scale benchmarks (due "soon").

The new Config object means we need to deprecate the
`DisablePodsFingerprinting` bool we added in the 4.11 cycle.
This also plays nice with the ongoing deprecation of the bool
flags in the k8s APIs.

Signed-off-by: Francesco Romani <fromani@redhat.com>