rte: sched: manifest: enable sched plugin tracing
Create empty directory using tmpfs to hold the NRT cache traces
if they are enabled; if not this is not to consume relevant resources,
so we can keep it always enabled.

Note this is a NOP at this point in time - the scheduler plugin
code is still under development. But the above reasoning still holds:
from both operator and scheduler plugin this is a very cheap and
backward compatible change.

Enable PFP debugging for  RTEs, consuming the functionality from the
latest u/s release. Enable dumping on their PFP status files on tmpfs.
Even more than for the scheduler plugin, we can keep this always enabled
from storage perspective.

Signed-off-by: Francesco Romani <fromani@redhat.com>