test: internal: dissolve schedutils
Package names like `utils` are a antipattern
(https://dave.cheney.net/2019/01/08/avoid-package-names-like-base-util-or-common)

in this patch we dissolve 'schedutils' in `internal/nodes` and (mostly)
in the new package `podlist`, which wants to list all the pods belonging
to a deployment/replicaset/daemonset.

Arguably, `podlist` is not the greatest name, and probably the package
content would desertve more thoughts, but it's still a step in the right
direction farther from `utils`, and will enable code reuse and better
testing.

Signed-off-by: Francesco Romani <fromani@redhat.com>