rte: add exclude list for pod namespaces/name
Enable the RTE to filter out pods using namespace/name globbing (not
regex) when listing through podresouces API.

The main driver for this functionality is to handle lingering
terminated pods which would not be detected by the scheduler plugin,
and that would make the pod fingerprint computation drift, thus
stalling the scheduler.
Filtering is optional and disabled by default.

Enable the handling of the filter list from the operator.

Signed-off-by: Francesco Romani <fromani@redhat.com>