serial: update scheduler name properly
The test fails to set the scheduler name under the test's pod because nroSchedObj.Spec.SchedulerName does not exist, thus it returns an empty value, which later would be translated to the default kube scheduler name instead of the test scheduler name.
The scheduler name is defined under Status not under Spec of the numaresourcesscheduler CR.

Modify the scheduler name properly in Status.SchedulerName to check the anticipated test's result.

get it from status instead of spec.

Signed-off-by: shereenH <shajmakh@redhat.com>