controllers: add and use the normalize package
We need to normalize the NodeGroupConfig earlier, because
we want to use the data to set the statuses and, in general,
because it makes it easier to work this way.

Normalization here means instianting all the default values
to be explicit, rather then implied.

We defer this logic to a new helper package.

Signed-off-by: Francesco Romani <fromani@redhat.com>