utils: skip accounting for CPUs and memory if pod is not gu
Till now, verifying the consumed resources in NRT was done for all resources passed as the resource list of the pod. That is not right when the pod is not guaranteed, emphasizing a burstable pod where it has at least one of the two resource types: CPU and memory.

By NRT design, only if the pod is guaranteed the CPU and memory amounts will be reflected in NRT as consumed under the numa zones, in addition to the other available resource types. If the pod is not guaranteed we have the following cases:
1. best effort: meaning no CPUs/memory is specified, but other resources might be requested and will be accounted for in NRT, thus no change is needed in this case.
2. burstable: at least one of CPUs/memory is specified, but because of the NRT design those resources are consumed from the shared pool and will not be reflected in NRT. OTOH other resource types will be reflected as consumed in NRT, thus we should continue accounting for them.

In this PR we are skipping accounting for the CPU and memory amount among the consumed amount of resources in NRT in case we are dealing with non-gu pod. The reason for adding this change now is because till now there was no test that was expected to schedule a burstable pod with resources that are not only CPU/memory. The next commit will add such test cases.

Signed-off-by: shereenH <shajmakh@redhat.com>