serial: support test for pod TMscope
The test with its older version is supporting only TMscope=container, because it creates the test's pods with two containers each expected to be placed on different numa zones of the target node. When running this scenario oc a cluster with tmscope=pod the test's pod with TAS scheduler fails to run because of insufficient resource availability on a single numa node, while the pod with default scheduler hits TAE.

Support this TC to run on TMscope=pod by padding the target node zones leaving room only for the one pod resource and adjusting the pod.Spec accordingly.

Signed-off-by: shereenH <shajmakh@redhat.com>