serial: modify test to verify expected behavior
The test's goal is to verify that when scheduling two pods, the first with TAS scheduler and immediately after that a second pod with default scheduler, the first pod should go running and the second pod should keep pending, IOW the tests verify that TAS is not slow enough to let the default scheduler govern, thus verifying a framework prioritization is done properly for TAS scheduler. The test with its current version is waiting for the first pod to get running and only then it creates the second pod. Hence, the main test's purpose is not checked.

Modify the test in such a way that it will create both pods one after the other without waiting on their phases in between.

Signed-off-by: shereenH <shajmakh@redhat.com>