serial: verify pod is deleted
The current version of the test checks the NRT is restored directly after issuing the deletion command, this may be misleading in case of failure in this check because the root cause might be that the pod is not terminated yet.

Before moving to the next checks, verify that the pod is deleted and expect this operation to be successful.

Signed-off-by: shereenH <shajmakh@redhat.com>