serial: get expected target node
The test assumes the test's pods are placed on the target node, but that is not correct, eventhough all the non-target nodes are padded, something might go wrong and a pod could be scheduled on a non-target node.

Fetch the anticipated target node as teh only node that wasn't padded in the test's setup.

Signed-off-by: shereenH <shajmakh@redhat.com>