serial: add missing checks
Add missing checks for when the replicaset is created, and wait for it to be up and running before moving on with the next steps.
Similarly, when deleting the test workload, verify the pods are completely gone before moving forward.
Another missing check is verifying the pods are scheduled on the anticipated target node.

Signed-off-by: shereenH <shajmakh@redhat.com>