test: e2e: read device-related env vars once
Helpers to get the device names can be invoked freely during
the suite execution, which can lead to a proliferation
of redundant error message.

To reduce noise in the logs, we read the env vars once at module
init time and we just return these values when asked.

As beneficial side effect, we ensure value consistency. Albeit possible,
changing these values while the suite is running is very likely
to wreck havoc, so let's just prevent this.

Signed-off-by: Francesco Romani <fromani@redhat.com>