serial: update schedulerName via Spec
The test failed to update the scheduler name to a new value because it was updating the SchedulerName under Status, while in general, updating a field in k8s object should be done via Spec and the active value of the fields should be fetched from Status.
Fix the test so it modifies the field via Spec and fetches the active value from Status.

Signed-off-by: shereenH <shajmakh@redhat.com>