serial: wait for pods creation
Tests that create a resource like deployment/daemonset directly move forward with the next checks without verifyng whether the creation of the resource was done properly, a thing that can cause a failure when performing checks on pods under these resources because there was not enough time between creation of the resource and waiting for its pods to be created.

Add a verification after creating such resources that thier pods are created (NOT guaranteed to be scheduled/running) before moving on with the test.

For example, after adding this verification, we can see that at the very first moments of creating the deployment, it has 0 replicas, and this causes the next check to fail because it performs checks on these replicas:

```
W1109 18:07:04.070981  103558 workload_unschedulable.go:686] Waiting for deployment: "e2e-test-workload-unschedulable-8605/test-dp-47615" to have 6 replicas, current number of replicas: 0
```

Signed-off-by: shereenH <shajmakh@redhat.com>