serial: check pod events in a loop
The test may sometimes fail to find the TAE event the pod's events due to inaccurate timing as using the default scheduler will schedule the pod and the TAE event may take a moment to appear.
Enhance this check by verifying that eventually the pod has TAE event by checking the events every second up to a minute.

Signed-off-by: shereenH <shajmakh@redhat.com>