api: avoid maps in the API
The `PodExcludes` feature was implemented using a map,
but we should avoid map in APIs. So let's use a slice
of NamespacedName instead, which is still perfectly
fine for our goals.

Note there is not to date a Status representation of
PodExcludes, which is out of scope of this PR.

Signed-off-by: Francesco Romani <fromani@redhat.com>