machineconfigpools -> helper/nodegroup
A very sizeable chunk of the functions in `machineconfigpools`
are about nodegroup trees, and are strictly coupled with API objects.
In otder to make the coupling obvious, improve the layout and adhere
more to the k8s conventions, extract this subset (with their test)
and move them into a `helper` subpackage of the apis.

This makes much easier introducing new API versions in the future.

Signed-off-by: Francesco Romani <fromani@redhat.com>