API: introduce and wire the v1 package
the v1 API is 99% v1alpha1, but with the deprecated field
in NodeGroup gone and cleaned up. Add the APIs and wire
up them.

Note about webhooks:

If we trivially add the hook wiring code into the api/ package
per kubebuilder/operator-sdk defaults, we will have test failures
because the `-kubeconfig` flag will be redefined

The culprits are the k8s e2e test framework and controller-runtime.
The latter is imported because of api v1 webhooks.
the api v1 package is _ALSO_ imported in the v1alpha1 package because
of the conversion

Signed-off-by: Francesco Romani <fromani@redhat.com>