api: layout: api/numaresourcesoperator/v* -> api/v*
trivial path rename to comply with layout standards
No changes in content or behaviour, trivial
update change.

We don't have yet any consumer of the operator API
(nor we probably should), so this is still a good
time to do this invasive (albeit simple) change.

+++

why did we had the `api/numaresources/v1alpha` layout in the first
place?
The exact reason is lost in the beginning of the repo, but as far
as the original initiator (ffromani) memory goes, this was never
a conscious design decision, rather an artifact of how
`operator-sdk` was used (possibly even with some unusual but legal
parameter combinarion).

Further hints that the current layout is historical artifact can
be seen in the layout of the spiritual ancestor of
numaresources-operator:

https://github.com/openshift-kni/rte-operator/commit/3570aba082154d3dbf99f6947f484365dd58c58a

Signed-off-by: Francesco Romani <fromani@redhat.com>