e2e: serial: fix expected event on error
Catch up with u/s https://github.com/kubernetes-sigs/scheduler-plugins/pull/547

We fix the expected message within events when a container/pod is
unschedulable. Previously, we had to use custom strings, but now
we can leverage the constants declared in the `nrosched` utility
package, on which we depend already anyway.

Signed-off-by: Francesco Romani <fromani@redhat.com>