serial(v1beta1): set expected scheduler to empty for default sched
A test that verifies that a pod scheduled with default scheduler failed because the expected Event.ReportingController is not "default-scheduler".
Looking into the issue, it turned out that Event.ReportingController is set only if the scheduler is non-default starting v1beta1, if it is default the field is left empty.

Adopt this update in the test and pass an empty value as the expected scheduler.

Signed-off-by: shereenH <shajmakh@redhat.com>