scheduler: controller: omitempty CacheResyncPeriodSeconds
CacheResyncPeriodSeconds equals 0 shiuld actually be omitted
by the generated config to ensure backward compatibility.

In theory this should be handled automatically by the API machinery.
In practice, the external versioned API uses correctly a pointer field
with the omitempty annotation; the internal unversioned API does not.
Because of this impedence, the zero field find its way in the generated
config. Which works as intended with recent enough plugins.

With old plugins, the unexpected fields leads to the scheduler plugin
crashing because can't properly decode the config.

Hence, to ensure the fullest backward compatibility, we go the extra
mile and we add compatibility code to forcefully remove the field
when set to zero.

Note that since its inception the scheduler plugin defaults
`CacheResyncPeriodSeconds` to 0 (= missing data is interpreted as zero).

Signed-off-by: Francesco Romani <fromani@redhat.com>