serial: unbind object variable from Eventually
The gaol is to update nroOperObj with the initial configuration and proceed with that restored object also outside Eventually.
Having nroOperObj inside eventually block prevents us from using the updated object, which later causes the test to wait for incorrect mcps to be updated.

Move nroOperObj outside eventually block to enable exposing the latest version of the object later in the test.

Signed-off-by: shereenH <shajmakh@redhat.com>