e2e: serial: resacct: improve and stabilize tests
Improve the debuggability of resacct tests
Stabilize the resacct tests by being more strict about
the cluster conditions to wait for.
Previously, we just checked for _A_ change in the NRT list,
which was very crude and caused frequent flakes.
Now we wait for a targeted change in the very NRT object we
care about, then we check that change is what we expected.
This is more precise, more correct and, at least in the
initial runs, less flaky.

In general, this seems a much better direction anyway,
so let's fix it.

Signed-off-by: Francesco Romani <fromani@redhat.com>