e2e: serial: resacct: reuse helpers
We should in general very, VERY careful when factorizing
code in e2e tests. We totally DO NOT want to add coupling
to tests, and when we factor code we need to make really sure
we're factoring code which does so general that a failure or
a bug _always_ means a critical failure in the context of tests.

Easy candidates are setup helpers and expectations.

Turns out we check quite some expectations about proper
NRT resource accounting, so let's reuse expectations for the
most obvious candidate sites.

Signed-off-by: Francesco Romani <fromani@redhat.com>