internal: wait: add helper to wait for stable NRT
In quite a few places we want to check NRT values.
In order to do meaningful checks we need however to make sure
that NRT data caught up with the changes made to the cluster.

This is a nontrivial problem in the current architecture.
To move in the right direction, we start adding helpers to check
the NRT data is "stable". We call NRT data "stable" when it is not
changing "for a while", and "for a while" is few consecutive update
interval. This is a pretty crude approximation but it's the best
we can do atm.

Signed-off-by: Francesco Romani <fromani@redhat.com>