e2e: utils: fix nrt.GetUpdated
Fix the `GetUpdated` helper. Previously we naively
used the ResourceVersion as gate to detect changes. This is wrong.
If ResourceVersion is equal, it means the object is not changed (yet).
But if it is different, it really means nothing and we need to check
if the fields we care about (vs metadata, status in general...) changed
and if so, react accordingly.

This patch fixes this bug, and likely quite some flakes.

Signed-off-by: Francesco Romani <fromani@redhat.com>